refactor(core): P5 - add formatProcessError helper and normalize error patterns
- Add formatProcessError(result, fallback) to process.ts
- Replace inline stderr/stdout fallback triples in 10 feature files
- Remaining P4 import migrations bundled in this commit
